Creates modules and improvements at code.
