+++ .idea/gradle.xml
@@ -9,6 +9,8 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/example" />
+            <option value="$PROJECT_DIR$/sync" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
