+++ app/src/main/java/com/oxeanbits/forecastchart/model/Line.kt
@@ -2,7 +2,7 @@ package com.oxeanbits.forecastchart.model
 
 import com.github.mikephil.charting.data.Entry
 
-data class LineDataCreator (
+data class Line (
     val values: ArrayList<Entry>,
     val label: String,
     val color: Int,
