+++ settings.gradle
@@ -1,2 +1,2 @@
-include ':app'
+include ':app', ':example', ':sync'
 rootProject.name='ForecastChart'
