+++ sync/src/main/res/layout/marker_layout.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/black">
+    android:background="#000000">
 
     <TextView
         android:id="@+id/prod_content"
@@ -12,7 +12,7 @@
         android:paddingRight="5dp"
         android:textSize="13dp"
         android:text="7m³"
-        android:textColor="@color/white"
+        android:textColor="#FFFFFF"
         android:textStyle="bold"
         />
     <TextView
@@ -23,7 +23,7 @@
         android:paddingRight="5dp"
         android:textSize="11dp"
         android:text="2019-01-01"
-        android:textColor="@color/white"
+        android:textColor="#FFFFFF"
         android:textStyle="bold"
         />
